Organize src folder.
Now we only export luaopen_luv.  This should reduce the area of conflict
when importing into luvit eventually.

The source files are now broken up by responsibility, but all act as a
single file thanks to C includes

There are now Makefiles for lua, luajit and luvit.
