+++ src/util.c
@@ -14,9 +14,54 @@
  *  limitations under the License.
  *
  */
-#include <stdlib.h>
-#include <assert.h>
-#include "common.h"
+
+/* Unique type codes for each uv type */
+enum luv_type {
+  LUV_HANDLE  = 0x01,
+  LUV_TIMER   = 0x02,
+  LUV_STREAM  = 0x04,
+  LUV_TCP     = 0x08,
+  LUV_UDP     = 0x10,
+  LUV_TTY     = 0x20,
+  LUV_PIPE    = 0x30,
+  LUV_PROCESS = 0x80
+};
+
+/* Mask is the types that can be extracted from this concrete type */
+enum luv_mask {
+  LUV_TIMER_MASK    = LUV_HANDLE | LUV_TIMER,
+  LUV_TCP_MASK      = LUV_HANDLE | LUV_STREAM | LUV_TCP,
+  LUV_UDP_MASK      = LUV_HANDLE | LUV_STREAM | LUV_UDP,
+  LUV_TTY_MASK      = LUV_HANDLE | LUV_STREAM | LUV_TTY,
+  LUV_PIPE_MASK     = LUV_HANDLE | LUV_STREAM | LUV_PIPE,
+  LUV_PROCESS_MASK  = LUV_HANDLE | LUV_PROCESS
+};
+
+/* luv handles are used as the userdata type that points to uv handles.
+ * The luv handle is considered strong when it's "active" or has non-zero
+ * reqCount.  When this happens ref will contain a luaL_ref to the userdata.
+ */
+typedef struct {
+  uv_handle_t* handle; /* The actual uv handle. memory managed by luv */
+  int refCount;        /* a count of all pending request to know strength */
+  lua_State* L;        /* L and ref together form a reference to the userdata */
+  int threadref;       /* hold reference to coroutine if created in one */
+  int ref;             /* ref is null when refCount is 0 meaning we're weak */
+  int mask;
+} luv_handle_t;
+
+typedef struct {
+  luv_handle_t* lhandle;
+  int data_ref;
+  int callback_ref;
+} luv_req_t;
+
+typedef struct {
+  lua_State* L;
+  int ref;
+} luv_callback_t;
+
+static lua_State* luv_main_thread;
 
 /* Initialize a new lhandle and push the new userdata on the stack. */
 static luv_handle_t* luv_handle_create(lua_State* L, size_t size, int mask) {
@@ -50,32 +95,32 @@ static luv_handle_t* luv_handle_create(lua_State* L, size_t size, int mask) {
   return lhandle;
 }
 
-uv_timer_t* luv_create_timer(lua_State* L) {
+static uv_timer_t* luv_create_timer(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_timer_t), LUV_TIMER_MASK);
   return (uv_timer_t*)lhandle->handle;
 }
 
-uv_tcp_t* luv_create_tcp(lua_State* L) {
+static uv_tcp_t* luv_create_tcp(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_tcp_t), LUV_TCP_MASK);
   return (uv_tcp_t*)lhandle->handle;
 }
 
-uv_udp_t* luv_create_udp(lua_State* L) {
+static uv_udp_t* luv_create_udp(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_udp_t), LUV_UDP_MASK);
   return (uv_udp_t*)lhandle->handle;
 }
 
-uv_tty_t* luv_create_tty(lua_State* L) {
+static uv_tty_t* luv_create_tty(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_tty_t), LUV_TTY_MASK);
   return (uv_tty_t*)lhandle->handle;
 }
 
-uv_pipe_t* luv_create_pipe(lua_State* L) {
+static uv_pipe_t* luv_create_pipe(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_pipe_t), LUV_PIPE_MASK);
   return (uv_pipe_t*)lhandle->handle;
 }
 
-uv_process_t* luv_create_process(lua_State* L) {
+static uv_process_t* luv_create_process(lua_State* L) {
   luv_handle_t* lhandle = luv_handle_create(L, sizeof(uv_process_t), LUV_PROCESS_MASK);
   return (uv_process_t*)lhandle->handle;
 }
@@ -90,48 +135,48 @@ static luv_handle_t* luv_get_lhandle(lua_State* L, int index, int type) {
   return lhandle;
 }
 
-uv_handle_t* luv_get_handle(lua_State* L, int index) {
+static uv_handle_t* luv_get_handle(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_HANDLE);
   return (uv_handle_t*)lhandle->handle;
 }
 
-uv_timer_t* luv_get_timer(lua_State* L, int index) {
+static uv_timer_t* luv_get_timer(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_TIMER);
   return (uv_timer_t*)lhandle->handle;
 }
 
-uv_stream_t* luv_get_stream(lua_State* L, int index) {
+static uv_stream_t* luv_get_stream(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_STREAM);
   return (uv_stream_t*)lhandle->handle;
 }
 
-uv_tcp_t* luv_get_tcp(lua_State* L, int index) {
+static uv_tcp_t* luv_get_tcp(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_TCP);
   return (uv_tcp_t*)lhandle->handle;
 }
 
-uv_udp_t* luv_get_udp(lua_State* L, int index) {
-  luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_UDP);
-  return (uv_udp_t*)lhandle->handle;
-}
+// static uv_udp_t* luv_get_udp(lua_State* L, int index) {
+//   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_UDP);
+//   return (uv_udp_t*)lhandle->handle;
+// }
 
-uv_tty_t* luv_get_tty(lua_State* L, int index) {
+static uv_tty_t* luv_get_tty(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_TTY);
   return (uv_tty_t*)lhandle->handle;
 }
 
-uv_pipe_t* luv_get_pipe(lua_State* L, int index) {
+static uv_pipe_t* luv_get_pipe(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_PIPE);
   return (uv_pipe_t*)lhandle->handle;
 }
 
-uv_process_t* luv_get_process(lua_State* L, int index) {
+static uv_process_t* luv_get_process(lua_State* L, int index) {
   luv_handle_t* lhandle = luv_get_lhandle(L, index, LUV_PROCESS);
   return (uv_process_t*)lhandle->handle;
 }
 
 /* This needs to be called when an async function is started on a lhandle. */
-void luv_handle_ref(lua_State* L, luv_handle_t* lhandle, int index) {
+static void luv_handle_ref(lua_State* L, luv_handle_t* lhandle, int index) {
   /* If it's inactive, store a ref. */
   if (!lhandle->refCount) {
     lua_pushvalue(L, index);
@@ -141,7 +186,7 @@ void luv_handle_ref(lua_State* L, luv_handle_t* lhandle, int index) {
 }
 
 /* This needs to be called when an async callback fires on a lhandle. */
-void luv_handle_unref(lua_State* L, luv_handle_t* lhandle) {
+static void luv_handle_unref(lua_State* L, luv_handle_t* lhandle) {
   lhandle->refCount--;
   assert(lhandle->refCount >= 0);
   /* If it's now inactive, clear the ref */
@@ -156,7 +201,7 @@ void luv_handle_unref(lua_State* L, luv_handle_t* lhandle) {
 }
 
 /* Get L from a lhandle (moving to the main thread if needed) and push the userdata on the stack. */
-lua_State* luv_prepare_event(luv_handle_t* lhandle) {
+static lua_State* luv_prepare_event(luv_handle_t* lhandle) {
   lua_State* L = lhandle->L;
 #ifdef LUV_STACK_CHECK
   int top = lua_gettop(L);
@@ -173,7 +218,7 @@ lua_State* luv_prepare_event(luv_handle_t* lhandle) {
 /* Get a named callback.  If it's there, push the function and the userdata on the stack.
    otherwise return 0
    either way, the original userdata is removed. */
-int luv_get_callback(lua_State* L, const char* name) {
+static int luv_get_callback(lua_State* L, const char* name) {
   int isfunc;
 #ifdef LUV_STACK_CHECK
   int top = lua_gettop(L);
@@ -198,7 +243,7 @@ int luv_get_callback(lua_State* L, const char* name) {
 
 /* returns the lua_State* and pushes the callback onto the stack */
 /* STACK +1 */
-lua_State* luv_prepare_callback(luv_req_t* lreq) {
+static lua_State* luv_prepare_callback(luv_req_t* lreq) {
   luv_handle_t* lhandle = lreq->lhandle;
   lua_State* L = lhandle->L;
 #ifdef LUV_STACK_CHECK
@@ -213,7 +258,7 @@ lua_State* luv_prepare_callback(luv_req_t* lreq) {
   return L;
 }
 
-void luv_call(lua_State *C, int nargs, int nresults) {
+static void luv_call(lua_State *C, int nargs, int nresults) {
   lua_State* L = luv_main_thread;
   if (L != C) {
     /* Move the function call to the main thread if it's in a coroutine */
@@ -222,40 +267,100 @@ void luv_call(lua_State *C, int nargs, int nresults) {
   lua_call(L, nargs, nresults);
 }
 
-void luv_stack_dump(lua_State* L, int top, const char* name) {
-  int i, l;
-  printf("\nAPI STACK DUMP: %s\n", name);
-  for (i = top, l = lua_gettop(L); i <= l; i++) {
-    const char* typename = lua_typename(L, lua_type(L, i));
-    switch (lua_type(L, i)) {
-      case LUA_TNIL:
-        printf("  %d: %s\n", i, typename);
-        break;
-      case LUA_TNUMBER:
-        printf("  %d: %s\t%f\n", i, typename, lua_tonumber(L, i));
-        break;
-      case LUA_TBOOLEAN:
-        printf("  %d: %s\n\t%s", i, typename, lua_toboolean(L, i) ? "true" : "false");
-        break;
-      case LUA_TSTRING:
-        printf("  %d: %s\t%s\n", i, typename, lua_tostring(L, i));
-        break;
-      case LUA_TTABLE:
-        printf("  %d: %s\n", i, typename);
-        break;
-      case LUA_TFUNCTION:
-        printf("  %d: %s\t%p\n", i, typename, lua_tocfunction(L, i));
-        break;
-      case LUA_TUSERDATA:
-        printf("  %d: %s\t%p\n", i, typename, lua_touserdata(L, i));
-        break;
-      case LUA_TTHREAD:
-        printf("  %d: %s\t%p\n", i, typename, lua_tothread(L, i));
-        break;
-      case LUA_TLIGHTUSERDATA:
-        printf("  %d: %s\t%p\n", i, typename, lua_touserdata(L, i));
-        break;
-    }
+// Commented out because it's unused normally.
+// static void luv_stack_dump(lua_State* L, int top, const char* name) {
+//   int i, l;
+//   printf("\nAPI STACK DUMP: %s\n", name);
+//   for (i = top, l = lua_gettop(L); i <= l; i++) {
+//     const char* typename = lua_typename(L, lua_type(L, i));
+//     switch (lua_type(L, i)) {
+//       case LUA_TNIL:
+//         printf("  %d: %s\n", i, typename);
+//         break;
+//       case LUA_TNUMBER:
+//         printf("  %d: %s\t%f\n", i, typename, lua_tonumber(L, i));
+//         break;
+//       case LUA_TBOOLEAN:
+//         printf("  %d: %s\n\t%s", i, typename, lua_toboolean(L, i) ? "true" : "false");
+//         break;
+//       case LUA_TSTRING:
+//         printf("  %d: %s\t%s\n", i, typename, lua_tostring(L, i));
+//         break;
+//       case LUA_TTABLE:
+//         printf("  %d: %s\n", i, typename);
+//         break;
+//       case LUA_TFUNCTION:
+//         printf("  %d: %s\t%p\n", i, typename, lua_tocfunction(L, i));
+//         break;
+//       case LUA_TUSERDATA:
+//         printf("  %d: %s\t%p\n", i, typename, lua_touserdata(L, i));
+//         break;
+//       case LUA_TTHREAD:
+//         printf("  %d: %s\t%p\n", i, typename, lua_tothread(L, i));
+//         break;
+//       case LUA_TLIGHTUSERDATA:
+//         printf("  %d: %s\t%p\n", i, typename, lua_touserdata(L, i));
+//         break;
+//     }
+//   }
+//   printf("\n");
+// }
+
+static const char* type_to_string(uv_handle_type type) {
+  switch (type) {
+#define XX(uc, lc) case UV_##uc: return ""#lc"";
+    UV_HANDLE_TYPE_MAP(XX)
+#undef XX
+    case UV_UNKNOWN_HANDLE: return "unknown";
+    case UV_FILE: return "file";
+    default: return "";
   }
-  printf("\n");
+}
+
+static int new_tcp(lua_State* L) {
+  uv_tcp_t* handle = luv_create_tcp(L);
+  if (uv_tcp_init(uv_default_loop(), handle)) {
+    uv_err_t err = uv_last_error(uv_default_loop());
+    return luaL_error(L, "new_tcp: %s", uv_strerror(err));
+  }
+  return 1;
+}
+
+static int new_udp(lua_State* L) {
+  uv_udp_t* handle = luv_create_udp(L);
+  if (uv_udp_init(uv_default_loop(), handle)) {
+    uv_err_t err = uv_last_error(uv_default_loop());
+    return luaL_error(L, "new_udp: %s", uv_strerror(err));
+  }
+  return 1;
+}
+
+static int new_timer(lua_State* L) {
+  uv_timer_t* handle = luv_create_timer(L);
+  if (uv_timer_init(uv_default_loop(), handle)) {
+    uv_err_t err = uv_last_error(uv_default_loop());
+    return luaL_error(L, "new_timer: %s", uv_strerror(err));
+  }
+  return 1;
+}
+
+static int new_tty(lua_State* L) {
+  uv_tty_t* handle = luv_create_tty(L);
+  uv_file fd = luaL_checkint(L, 1);
+  int readable = lua_toboolean(L, 2);
+  if (uv_tty_init(uv_default_loop(), handle, fd, readable)) {
+    uv_err_t err = uv_last_error(uv_default_loop());
+    return luaL_error(L, "new_tty: %s", uv_strerror(err));
+  }
+  return 1;
+}
+
+static int new_pipe(lua_State* L) {
+  uv_pipe_t* handle = luv_create_pipe(L);
+  int ipc = lua_toboolean(L, 1);
+  if (uv_pipe_init(uv_default_loop(), handle, ipc)) {
+    uv_err_t err = uv_last_error(uv_default_loop());
+    return luaL_error(L, "new_pipe: %s", uv_strerror(err));
+  }
+  return 1;
 }
